open samples update
This is to close #43.

The text around the open samples was adjusted and now says "may" as opposed to in such cases, so this should be clear that it is not necessary at all.

I removed the sentence about the location of the testing. I think this is something the vendor, lab and scheme will need to work out for an evaluation and putting it here as a statement up front clearly causes a lot of questions about whether this is sufficient or not. Given that the iTC cannot answer the question about it being sufficient adequately for all evaluations and circumstances, I think it is best to remove that sentence and just be clear about the purposes of this test environment, which is really about the samples being used, and should not be about where they are located.

The point about whether all devices need to be tested is handled in Appendix B, and so does not need to be noted here. A note about looking there could be added, but I do not think this is relevant for consideration under the topic here (bringing that in only confuses the reader, like the location does).

Lastly, I checked several other iTC SDs and do not see anything in them about that the scheme can accept or not. I think this is just expected based on endorsements and general use as to what a scheme will accept, I do not thing we should call it out here.