Changes to update FI testing
This is to resolve #74. The idea with this change is that all the tests are actually fault injection (since that is what they are) and to point the FI testing to be done as part of FPT_PHP.3 since that incorporates the "how to induce the fault" part. This could technically be done the other way (test FPT_PHP.3 as part of FPT_FLS.1/FI), but I don't think that really matters as long as we are testing appropriately.