side-channel
This removes the specific reference to the side channel attacks. It also adjusted the intro to be more explicit about this being targeted to embedded components/systems in a device, and not discrete components like TPMs, HSMs or secure elements which may be included in a device, but are not technically embedded in the same way.

I do not think we have a problem stating this, and if someone really wants to complain about removing secure elements or TPMs from this they can speak up during comments and we can adjust, but I doubt that will happen as this is targeting the middle part between them.

One thing here is that it may be useful to add the definition for Secure Execution Environment that has been defined in the Biometrics PP-Module as a generic term for things like a TEE but which are focused on the idea it is separated from the main OS, whatever it may be.