FCS_CKM.2 encrypted channels
The catalog FCS_CKM.2 provides "encrypted channels" as an option for the way to transport keys. This has a dependency on FTP_PRO.1, which we don't want to bring in (this was already discussed as a possible edit for CC:2022 and rejected in #148).

The 1.0 has defined FTP_ITP_EXT.1, FTP_ITE_EXT.1 and FTP_ITC_EXT.1 as transport methods so I have adjusted the SFR to point to those methods used for SDEs/SDOs as acceptable key transport methods.

I'm actually not sure we need anything other than the FTP_IPT_EXT.1 for physically protected channels, but this is more complete.