key integrity
This is to close #3.

Instead of adding the key wrapping, I adjusted the text to just state it protects integrity and to reference the appropriate objective. This is the only place in the intro (at least this section) where SFRs were actually referenced, so it seemed out of place to have them referenced here and to add KW or GCM.