adding "OS" to DSC description
This is to close #4. This adds "OS or" to match with the description of the SD. I think this is needed since some DSCs may encompass some sort of OS (like a TEE) that is more than just firmware. This may be a meaningless distinction and OS could be as easily removed from the SD, but when I think firmware I don't tend to think of larger systems like a TEE but smaller ones like on an eSE.