simplify fia_uau.6
This is to close #219 by simplifying the selection to no choice.

As the app note states, the specific options allowed are set in FMT_MSA.3, so also setting them here adds confusion. All the options are specified according toe SDO.Reauth, so there isn't anything else that has to be specified, one of those settings must be used.