RoT Test updates
This is to close #280.

The changes includes several of the RoT requirements to make sure the intent is covered. I adjusted the FPT_PRO_EXT.1 to move some of the TSS stuff into the KMD as I think that is a better place for that info (I can't imagine details of the RoT being public).

I modified the language for the testing to be consistent with some other EAs which sounds better.

I added FDP_SDI.2 for FPT_ROT_EXT.2 as that is mentioned in the app note and is relevant.