Crypto WG update
This is the update to incorporate the updated changes from the Crypto WG.

Consider if FCS_CKM_EXT.9 should be included as this is new. It seems like it could be relevant, but it isn't something that is included at the moment, and it isn't clear this is necessary at this time (it could just be added as an optional requirement).