Addition of Direct Rationale
As specified in https://github.com/DSC-iTC/cPP/issues/35 the approach of
the direct rationale PP/ST shall be taken. The change removes the
objectives for the TOE and replaces the objectives to SFR mapping with
the SPD to SFR mapping as defined in CC2022 Appendix Part 1 D.4.

The approach to the mapping was to apply the threads mapped to
objectives when replacing the objectives in the SFR mapping table.