reference update
This updates the references in the document to the latest iterations of the specs that are published. The catalog doesn't have some of these, but it would seem that these are the right ones to list given they are the current versions and the older ones are no longer considered up to date.

This is to close #109