updates to the dependency table
This is to close #347

Looking at the catalog, the following dependencies seem wrong (in the catalog, the table or both):

FCS_COP.1/KeyedHash - it would seem that FCS_COP.1/Hash is a dependency, since SHA is required for HMAC-SHA to work. Also, it lists FCS_CKM.5 as a dependency, but that seems backwards. Further, several of the other ones don't make a lot of sense.