FCS_CKM_EXT.3 hidden requirement
This is to close #370.

The logic here is that selecting the software storage does not explicitly mandate the use of FCS_CKM_EXT.3, but that is the intent. Having this as an app note is a hidden requirement and is not allowed. By adding in the statement that protection of software-based storage has to be according to that SFR, it makes that requirement visible and explicit.

With that I don't think we actually need to have the app note anymore, since it really doesn't have any reason to be there. We do have this in some other places, but when it is explicit, I think we can drop it.