FDP_DSC.2 EA
This is to close #372.

The point of this change is to be minimally invasive and to point to other tests that would actually verify the underlying activities to support this SFR. The only dependency listed in FCS_COP.1, but based on how the cPP is designed FPT_PHP.3 is needed as well (and FCS_STG_EXT.1.1 ties what is needed where, sort of, but the closest we have).